+++ indexer/feature_loader_base.hpp
@@ -37,7 +37,7 @@ namespace feature
     {
       return m_header.GetCodingParams(scaleIndex);
     }
-    inline int GetScalesCount() const { return m_header.GetScalesCount(); }
+    inline int GetScalesCount() const { return static_cast<int>(m_header.GetScalesCount()); }
     inline int GetScale(int i) const { return m_header.GetScale(i); }
   };
